First Commit!
I have decided to move my development  for my monopoly to Flutter/Dart in order to get a better understanding of the language and due to how many different platforms Flutter apps are created in.

So far, I have only gotten to the menus and will get to game creation soon.